refactor cli things to directory for organization
