+++ cli/__init__.py
@@ -1,0 +1,6 @@
+from .register import register
+#from .utils.env import load_environment_variables, load_readme
+from .identity import identity
+from .readme import readme
+
+__all__ = ['register', 'identity', 'readme']
